In imitatio, creatio ex nihilo
This is an initial attempt to recreate the behavior of the NodeJS-based
git-notify from Jani Eväkallio. I really liked the idea, but did not
like the dependence on NodeJS for something that felt like it could just
as easily be a Bash script. So I set out on a road to show that.

There is still much that needs to be done before I would call this
finished. In no particular order:

1. I would like to add some [bats-based][1] tests.
2. The centered notification messages from the original library are a
   nice-to-have.
3. Likewise with the hex colors for true color terminals, with sensible
   fallbacks.
4. Since this is a Git subcommand, it needs a manual page as well.
5. And probably a Makefile to build it for installation so it's easy to
   package.
6. A changelog for tracking changes over time.
7. A leightweight hook installer, even if it's a set of scripts to
   download.

None of these are particularly difficult on their own, but I have other
things on my mind so I'm putting this out into the world as-is.

[1]: https://github.com/sstephenson/bats